discord-bot/pyproject.toml
[tool.poetry]
name = "repl_python3_lazy-gambler-discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
twilio = "^7.3.2"
cleaner = "^0.4.0"
discord = "^1.7.3"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
maxleap-sdk = "^0.2.0"
nba-api = "^1.1.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
